Renamed DashboardView to TransactionsView
