+++ frontend/src/router/index.js
@@ -1,14 +1,14 @@
 import Vue from 'vue';
 import VueRouter from 'vue-router';
-import DashboardView from '../views/DashboardView.vue';
+import TransactionsView from '../views/TransactionsView.vue';
 
 Vue.use(VueRouter);
 
 const routes = [
   {
     path: '/',
-    name: 'dashboard',
-    component: DashboardView,
+    name: 'transactions',
+    component: TransactionsView,
   },
 ];
